updated Utils and ajdusted long to long long for MSVC

+++ week_2/day_11/intcode.h
@@ -24,11 +24,11 @@ private:
    void read_opcode();
    void execute_instruction();
    std::vector<int> mode;
-   long instruction;
+   long long instruction;
    bool running;
    bool pause;
    int  pos_in_memory;
-   long rb;
+   long long rb;
 
 
 public:
