+++ week_4/day_23/intcode.h
@@ -24,11 +24,11 @@ private:
    void read_opcode();
    void execute_instruction();
    std::vector<int> mode;
-   long instruction;
+   long long instruction;
    bool running;
    int  input_i;
    int  pos_in_memory;
-   long rb;
+   long long rb;
 
 
 public:
